Revert "feat: 홈화면, 홈화면 상세/정인"
