--- MagicSoraGodong/Models/Item.swift
@@ -1,20 +1,0 @@
-//
-//  Item.swift
-//  MagicSoraGodong
-//
-//  Created by 이정인 on 2020/10/21.
-//
-
-import Foundation
-
-class Item{
-    let itemTitle:String?
-    let itemPrice:String?
-    
-    init(title:String,price:String) {
-        self.itemTitle = title
-        self.itemPrice = price
-    }
-    
-    
-}
